changed all the arguments in the calls to the CPU_time function from double precision to single precision (which is supposed to be the standard --even though double precision seems to be working as well).

+++ Template/NLO/Source/alfas_functions.f
@@ -75,8 +75,8 @@ ccc
       external ALPHAS_not_timed
 
 c     timing statistics
-      double precision tbefore, tAfter
-      double precision tTot, tOLP, tFastJet, tPDF
+      real*4 tbefore, tAfter
+      real*4 tTot, tOLP, tFastJet, tPDF
       common/timings/tTot, tOLP, tFastJet, tPDF
 
 c     This function takes roughly 0.6 micro-seconds and the function
